# backend/openapi.yaml
openapi: 3.1.0
info:
  title: CadastraVision API
  description: "AI-enabled automated cadastral mapping platform. Programme: CadastralMap\
    \ \u2014 SIH 2026, PS 26012. Backend owner: Shiva."
  version: 0.2.0
paths:
...
  /v1/change-detection/run:
    post:
      tags:
      - Change Detection
      summary: Run change detection
      description: "Compare two imagery tiles to detect land-use changes. **STUB**\
        \ \u2014 Akshaya's pipeline (Phase 3)."
      operationId: run_change_detection_v1_change_detection_run_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangeDetectionRunRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeDetectionRunResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/change-detection/{job_id}:
    get:
      tags:
      - Change Detection
      summary: Get change detection job status
      description: "Poll the status of a change detection job. **STUB** \u2014 Phase\
        \ 3."
      operationId: get_change_detection_job_v1_change_detection__job_id__get
      parameters:
      - name: job_id
        in: path
        required: true
        schema:
          type: string
          title: Job Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeDetectionJobResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /v1/exports:
    post:
      tags:
      - Exports
      summary: Create export
      description: "Enqueue a data export (GeoJSON, Shapefile, GeoPackage, CSV). **STUB**\
        \ \u2014 Phase 3."
      operationId: create_export_v1_exports_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateExportRequest'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/exports/{export_id}:
    get:
      tags:
      - Exports
      summary: Get export status
      description: "Poll the status of an export job. **STUB** \u2014 Phase 3."
      operationId: get_export_status_v1_exports__export_id__get
      parameters:
      - name: export_id
        in: path
        required: true
        schema:
          type: string
          title: Export Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/audit/verify/{parcel_id}:
    get:
      tags:
      - Audit
      summary: Verify audit chain integrity
      description: Verify the SHA-256 hash-chain integrity for a parcel's audit trail.
      operationId: verify_audit_chain_v1_audit_verify__parcel_id__get
      parameters:
      - name: parcel_id
        in: path
        required: true
        schema:
          type: string
          title: Parcel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditVerifyResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/audit/export/{batch_id}:
    get:
      tags:
      - Audit
      summary: Export audit batch
      description: Export a batch of audit log entries.
      operationId: export_audit_batch_v1_audit_export__batch_id__get
      parameters:
      - name: batch_id
        in: path
        required: true
        schema:
          type: string
          title: Batch Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditExportResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /v1/audit/{parcel_id}:
    get:
      tags:
      - Audit
      summary: Get parcel audit trail
      description: Retrieve the complete audit trail for a parcel.
      operationId: get_parcel_audit_v1_audit__parcel_id__get
      parameters:
      - name: parcel_id
        in: path
        required: true
        schema:
          type: string
          title: Parcel Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditTrailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    AuditAction:
      type: string
      enum:
      - create
      - edit
      - approve
      - reject
      - sync
      - view
      - login
      - logout
      title: AuditAction
    AuditEntryResponse:
      properties:
        entry_id:
          type: string
          title: Entry Id
        entity_type:
          type: string
          title: Entity Type
        entity_id:
          type: string
          title: Entity Id
        action:
          $ref: '#/components/schemas/AuditAction'
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
        diff_json:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Diff Json
        ip_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Address
        created_at:
          type: string
          format: date-time
          title: Created At
        entry_hash:
          anyOf:
          - type: string
          - type: 'null'
          title: Entry Hash
          description: "SHA-256 hash of this entry (Prajith's chain \u2014 stub in\
            \ Phase 2)"
      type: object
      required:
      - entry_id
      - entity_type
      - entity_id
      - action
      - created_at
      title: AuditEntryResponse
    AuditExportResponse:
      properties:
        batch_id:
          type: string
          title: Batch Id
        download_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Download Url
        status:
          type: string
          title: Status
        entry_count:
          anyOf:
          - type: integer
          - type: 'null'
          title: Entry Count
      type: object
      required:
      - batch_id
      - status
      title: AuditExportResponse
      description: GET /v1/audit/export/{batch_id}.
    AuditTrailResponse:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        entries:
          items:
            $ref: '#/components/schemas/AuditEntryResponse'
          type: array
          title: Entries
        total:
          type: integer
          title: Total
      type: object
      required:
      - parcel_id
      - entries
      - total
      title: AuditTrailResponse
      description: GET /v1/audit/{parcel_id}.
    AuditVerifyResponse:
      properties:
        parcel_id:
          type: string
          title: Parcel Id
        chain_valid:
          type: boolean
          title: Chain Valid
        entry_count:
          type: integer
          title: Entry Count
        verification_note:
          type: string
          title: Verification Note
          default: 'Hash-chain verification: integration stub (Phase 2)'
      type: object
      required:
      - parcel_id
      - chain_valid
      - entry_count
      title: AuditVerifyResponse
      description: GET /v1/audit/verify/{parcel_id}.
...
    ChangeDetectionJobResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        status:
          type: string
          title: Status
        created_at:
          type: string
          format: date-time
          title: Created At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        changes_detected:
          anyOf:
          - type: integer
          - type: 'null'
          title: Changes Detected
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - job_id
      - status
      - created_at
      title: ChangeDetectionJobResponse
      description: GET /v1/change-detection/{job_id}.
    ChangeDetectionRunRequest:
      properties:
        tile_id_before:
          type: string
          title: Tile Id Before
          description: Baseline imagery tile ID
        tile_id_after:
          type: string
          title: Tile Id After
          description: New imagery tile ID to compare against
        area_of_interest_wkt:
          anyOf:
          - type: string
          - type: 'null'
          title: Area Of Interest Wkt
          description: Optional WKT polygon to restrict detection area
      type: object
      required:
      - tile_id_before
      - tile_id_after
      title: ChangeDetectionRunRequest
      description: POST /v1/change-detection/run.
    ChangeDetectionRunResponse:
      properties:
        job_id:
          type: string
          title: Job Id
        status:
          type: string
          title: Status
          default: queued
        message:
          type: string
          title: Message
          default: Change detection job queued.
      type: object
      required:
      - job_id
      title: ChangeDetectionRunResponse
...
    CreateExportRequest:
      properties:
        export_format:
          $ref: '#/components/schemas/ExportFormat'
          description: geojson | shapefile | geopackage | csv
        filters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Filters
          description: Optional filter params (zone, jurisdiction, status, etc.)
      type: object
      required:
      - export_format
      title: CreateExportRequest
      description: POST /v1/exports.
...
    ExportFormat:
      type: string
      enum:
      - geojson
      - shapefile
      - geopackage
      - csv
      title: ExportFormat
    ExportStatus:
      type: string
      enum:
      - queued
      - processing
      - complete
      - failed
      title: ExportStatus
    ExportStatusResponse:
      properties:
        export_id:
          type: string
          title: Export Id
        export_format:
          $ref: '#/components/schemas/ExportFormat'
        status:
          $ref: '#/components/schemas/ExportStatus'
        created_at:
          type: string
          format: date-time
          title: Created At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        download_url:
          anyOf:
          - type: string
          - type: 'null'
          title: Download Url
          description: "Signed download URL \u2014 available when status is complete"
        file_size_bytes:
          anyOf:
          - type: integer
          - type: 'null'
          title: File Size Bytes
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
      type: object
      required:
      - export_id
      - export_format
      - status
      - created_at
      title: ExportStatusResponse
      description: GET /v1/exports/{export_id}.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...